requirements.txt
asgiref
async-timeout
backports.zoneinfo;python_version<"3.9"
certifi
channels
channels-redis
charset-normalizer
cloudinary
django-cloudinary-storage
crispy-bootstrap4
crispy-tailwind
dj-database-url
Django
django-admin-interface
django-colorfield
django-crispy-forms
django-environ
django-jazzmin
django-phonenumbers
django-redis
django-storages
djangorestframework
geographiclib
geopy
googlemaps
gunicorn
idna
importlib-metadata
msgpack
packaging 
Pillow
psycopg2-binary
python-decouple
python-slugify
pytz
redis
requests
sqlparse
str2bool
tenacity
text-unidecode
typing_extensions
urllib3
websocket-client
websockets
whitenoise
zipp